requirements.txt
fastapi[standard]
uvicorn
requests
pydantic
python-dotenv
groq
google-generativeai
redis
